apps/api/pyproject.toml
[project]
name = "easyrag-api"
version = "0.1.0"
description = "easyRAG Medical RAG API — powered by RAG-Anything"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "raganything>=0.1.0",
    "lightrag-hku>=1.3.0",
    "pyjwt>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "aiofiles>=24.1.0",
    "pydantic-settings>=2.6.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
